Login Screen w/ SQLite Backend
Created a simple app that allows the user to create an account which has
both a username and password that are saved in the sqlite database. They
can then log in which will tell the user if the given password and
username are correct or not.

Next will be to make it so that when they log in, the system displays
the user home screen instead of just popping a dialog.